Move _move_glob to _file_glob
This better reflects its actual purpose
